add open_neovide_ alias and disable marvim and vim-smoothie
adds the `open_neovide_` alias to the `bash_aliases` file and disables the `marvim` and `vim-smoothie` plugins in the `vimrc` file. the `open_neovide_` alias provides an easy way to open neovide with arguments from the command line. the `marvim` and `vim-smoothie` plugins are disabled to improve the performance of vim.

the changes made to `bash_aliases` are:

```
+open_neovide_(){
+  if [ "$1" != "" ]
+  then
+    open /users/falcucci/neovide/target/release/bundle/osx/neovide.app --args $pwd/$1 --multigrid
+  else
+    open /users/falcucci/neovide/target/release/bundle/osx/neovide.app --args $pwd --multigrid
+  fi
+}
```

the changes made to `vimrc` are:

```
   plug 'chamindra/marvim'
-  plug 'psliwka/vim-smoothie'
+  " plug 'psliwka/vim-smoothie'
```