setup a reasonable leader and add git open aliases
adds a new alias for the git-open command to be used in the bash_aliases file. the alias `git-open` is mapped to the `gh browse` command which allows for easy opening of the current repository in the web browser.

in addition, this commit also adds the ability to copy and paste text in neovim. the `<d-v>` and `<d-c>` keys are mapped to the `+p` and `+y` commands respectively. this allows for the easy copy and paste of text in neovim.