increase smoothie speed and some shortcuts to toggle floaterm properly
increases the speed of the smoothie by changing the `smoothie_speed_constant_factor` and `smoothie_speed_linear_factor` to `40`, and adding a new `smoothie_speed_exponentiation_factor` set to `0.9`. additionally, the `floaterm_keymap_toggle` is set to `<f7>` and `copilot_enabled` is set to `1` to enable the copilot feature. finally, the `neovide_transparency` and `neovide_background_color` are set to `0.0` and `#000` respectively.

this commit changes the following files:
* `src/vim/vimrc`

the changes made are:
```diff
@@ -288,10 +288,12 @@ call plug#end()
 " make it interactive to turn on all my aliases
 " set shellcmdflag=-ic

-"  let g:smoothie_speed_constant_factor = 25
-"  let g:smoothie_speed_linear_factor = 25
+"  let g:smoothie_speed_constant_factor = 40
+"  let g:smoothie_speed_linear_factor = 40
+"  let g:smoothie_speed_exponentiation_factor = 0.9

 command float :floatermtoggle
+let g:floaterm_keymap_toggle = '<f7>'

 " adding copilot keybindings to turns it on and off
 " nnoremap <silent> <leader>cc :coccommand copilot.turnon<cr>
@@ -464,8 +466,6 @@ noremap <space> :
 nnoremap <f10> :tabp<cr>
 nnoremap <f12> :tabn<cr>

-nnoremap <f7> :gundotoggle<cr>
-
 " nnoremap <buffer> <f9> :exec '!clear;python' shellescape(@%, 1)<cr>

 " runs parsed files
@@ -598,7 +598,7 @@ map <leader>cc :tcomment<cr>

 " open a new float term
-map <leader>f :floatermtoggle<cr>
+map <leader>f :floatermnew<cr>

 " - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

@@ -880,8 +880,8 @@ if exists("g:neovide")
   set guifont=monaco:h20:b

   let g:neovide_transparency = 0.0
-  let g:transparency = 0.8
-  let g:neovide_background_color = '#0f1117'.printf('%x', float2nr(255 * g:transparency))
+  let g:transparency = 0.0
+  let g:neovide_background_color = '#000'
 "
   "remember previous window size
   let g:neovide_remember_window_size = v:true
@@ -889,7 +889,7 @@ if exists("g:neovide")
   let g:neovide_input_use_logo = v:true  " v:true on macos

   " enable copilot
-  let g:copilot_enabled = 0
+  let g:copilot_enabled = 1

   " allow copy paste in neovim
   let g:neovide_input_use_logo = 1
```

this commit increases the speed of the smoothie by changing the `smoothie_speed_constant_factor` and `smoothie_speed_linear_factor` to `40`, and adding a new `smoothie_speed_exponentiation_factor` set to `0.9`. additionally, the `floaterm_keymap_toggle` is set to `<f7>` and `copilot_enabled` is set to `1` to enable the copilot feature. finally, the `neovide_transparency` and `neovide_background_color` are set to `0.0` and `#000` respectively.

this commit changes the following file:
* `src/vim/vimrc`