add icat alias to bash_aliases and function to turn the neovide transparency mode
adds an alias for `icat` to `bash_aliases`. this alias will allow users to execute `kitty +kitten icat` from the terminal.

includes changes to the `bash_aliases` and `vimrc` files. for `bash_aliases`, the new alias `icat` was added on line 226. for `vimrc`, the value of `g:session_autosave` was changed from `yes` to `no` on line 390 and a new function `makeneovidetransparentandcharm` was added on lines 866-871. additionally, on line 908, a new mappping was added to allow users to call the `makeneovidetransparentandcharm` function with `f5`.