add commented goyo and limelight plugins as interesting plugings
adds the goyo and limelight plugins to our vimrc.

goyo is a plugin that adds a distraction-free writing mode to vim. it provides a clean, full screen writing environment that helps to remove distractions and increase focus.

limelight is a plugin that provides an even more minimal distraction-free writing mode. it dims everything except for the current line. this makes it easier to focus on the current line of text, and helps to keep you in the flow of writing.

diff --git a/src/vim/vimrc b/src/vim/vimrc
index 30c7ace037f6..9f23cf932b4c 100644
--- a/src/vim/vimrc
+++ b/src/vim/vimrc
@@ -278,6 +278,8 @@ plug 'skywind3000/asyncrun.vim'
 plug 'voldikss/vim-floaterm'
 plug 'voldikss/fzf-floaterm'
 plug 'mbbill/undotree'
+plug 'junegunn/goyo.vim'
+plug 'junegunn/limelight.vim'
 " plug 'psliwka/vim-smoothie'

 call plug#end()