theses changes are related to transparent and blur background configuration.
adds a few improvements to the vim configuration file. firstly, it sets the background of floaterm to black and makes neovide transparent. secondly, it adds keybindings for copying and pasting in neovide, and for tab navigation. finally, it adds a command for reloading the vim configuration file.

the changes made are as follows:
* added `autocmd vimenter,colorscheme * :hi floaterm guibg=black`
* added `autocmd vimenter * hi normal guibg=none ctermbg=none` and `set termguicolors`
* added `let g:transparency = 0.6` and `let g:neovide_background_color = '#0f1117'.printf('%x', float2nr(255 * g:transparency))`
* added `let g:neovide_input_use_logo = 1`
* added keybindings for copying and pasting in neovide, and for tab navigation
* added `command reload :source $myvimrc` and `noremap <f5> :reload <cr>`