add more shortcuts for navigating tabs
adds shortcuts for navigating tabs in vim, using `<d-j>` and `<d-k>` for tab-previous and tab-next respectively.

in `src/shell/bash_exports`, `/usr/local/bin/nvim` is added to the `$path` environment variable.

in `src/shell/bash_profile`, the `jbang` alias is added, and `$home/.jbang/bin` is added to `$path`.

in `src/shell/bashrc`, the `.bash-preexec.sh` script and the `atuin` command are removed from the `$path` environment variable.

in `src/vim/vimrc`, shortcuts for tab-previous and tab-next are added.