adds the `gruvbox8` theme to add a better light mode.
it adds the plugin `lifepillar/vim-gruvbox8` to the list of plugins and includes a new function to change the theme from dark to light, and vice versa.

the new function will map to the `f6` key, which will toggle the theme between dark and light. it also changes the background color of the neovide editor, the floaterm plugin, and the pmenu plugin. additionally, the `f4` key is remapped to call the `makeneovidetransparentandcharm` function.