change mappings for nerdtree and search
changes the mappings for nerdtree and search in `src/vim/vimrc`.

the mapping for nerdtree was changed from `ctrl+0` to `ctrl+e` and for search from `ctrl+r` to `enter`. in addition, the mapping for nerdtree was changed to include the directory of the current buffer when toggled.

the mapping for the emmet plugin was changed to `ctrl+e` and the option `g:neovide_hide_mouse_when_typing` was set to `v:false`.