update readme and vimrc and minor changes to transparent mode
updates the readme and vimrc files with some minor changes.

in readme, a new image is added and the alt text is updated.

in vimrc, some whitespaces are trimmed, `makeittransparent()` function is updated to make it slightly more transparent, and the floaterm guibg is updated to use a hex code instead of the color name.