add colorizer setup in nvim init-backup.lua
adds a setup for the nvim colorizer into the init-backup.lua file. this setup will enable users to toggle the colorizer for the current buffer with the `:colorizertoggle` command. the setup also includes a `loaded_colorizer` global to ensure the setup is only ran once.

the changes made to the init-backup.lua file are as follows:
- added `if vim.g.loaded_colorizer then return end`
- added `local command = vim.api.nvim_create_user_command`
- added `local c = require "colorizer"`
- added `c.setup()`
- added `command("colorizertoggle", function() ... end, {})`
- added `vim.g.loaded_colorizer = true`
- deleted `src/nvim/lua/nvim-bionic/init.lua`