add transparent mode picture in the readme
adds a new section to the readme for a transparent mode. this mode allows users to see the background of their operating system, and it can be used to customize the look and feel of the clipboard.

the commit includes an image to show what the output looks like on macos.

```diff
diff --git a/readme.md b/readme.md
index a20a3a8a0fde..28a34b269384 100644
--- a/readme.md
+++ b/readme.md
@@ -103,6 +103,22 @@ light mode:
     </tbody>
 </table>

+transparent mode:
+
+<table>
+    <tbody>
+        <tr>
+            <td>
+                <img src="https://raw.githubusercontent.com/falcucci/clipboard-images/main/images/cleanshot%202023-07-06%20at%2014.35.21%402x.png" alt="output for git status on macos" width="100%">
+            </td>
+        </tr>
+        <tr align="center">
+            <td>macos</td>
+        </td>
+    </tbody>
+</table>
+
+
 ## customize

 ### local settings
```